+++ Smith.MatrixSdk/MatrixApiConstants.cs
@@ -1,0 +1,7 @@
+﻿namespace Smith.MatrixSdk
+{
+    internal static class MatrixApiConstants
+    {
+        public const string LoginPasswordType = "m.login.password";
+    }
+}