+++ Smith.MatrixSdk/MatrixApiUris.cs
@@ -1,0 +1,7 @@
+﻿namespace Smith.MatrixSdk
+{
+    internal static class MatrixApiUris
+    {
+        public const string Login = "/_matrix/client/r0/login";
+    }
+}